Moved transceiver command mapping functions to commands.c
